Converting rome number to arabic
Converting rome numbers like "DCXI" to 611 up to 3999 arabic number ("MMMCMXCIX" rome). Output is -1 for invalid number.
I = 1; V = 5; X = 10; L = 50; C = 100; D = 500; M = 1000.
The idea is to compute patterns for all 4 ranks of output number and compare them with input string.
May be optimized soon, but this algorithm is very stable despite of a lot of manual array filling, comparements and repeating. Input string and arrays sizes are not so big.